Use subset function to reduce runing time
Using "for loop" to extract individual chromosome is time consuming especially when the .PEanno file is larger than 1G. (In my case, a 1.5GB files costs ~24 h)
To short the time, I apply "subset" function for extracting individual chromosome information, which is faster and consumes less resource. 
I have test the new code, and the output is 100% identical with the once run by previous code, but faster (<10 min for a 1.5GB files).